==> Notebooks/1_Data_split.ipynb <==
import pandas as pd 
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

# Data Load

# data_path = "/workspace/data/0_Subtask/[DICCD 분석] 251107.csv" # ~251110
data_path = "/workspace/data/0_Subtask/[DICCD 분석] 251111_SRD 추가.csv" # 251111
raw_df = pd.read_csv(data_path)


# test/ODD_CP 0 60개, ODD_CP 1 60개 추출
test_df_0 = raw_df[raw_df['ODD_CP'] == 0].sample(n=60, random_state=42)
test_df_1 = raw_df[raw_df['ODD_CP'] == 1].sample(n=60, random_state=42)
test_df = pd.concat([test_df_0, test_df_1])

# 나머지는 tune_df
tune_df = raw_df.drop(test_df.index).reset_index(drop=True)
import os 
# test, tune 분리된 결과 확인(선택)
len(test_df), len(tune_df)

result_dir = "/workspace/data/tune_data"
tune_df.to_csv(os.path.join(result_dir, "tune_df.csv"), index=False)
test_df.to_csv(os.path.join(result_dir, "test_df.csv"), index=False)

# 데이터 분리 결과 확인
print(f"Tune 데이터셋 크기: {len(tune_df)}")
print(f"Test 데이터셋 크기: {len(test_df)}")
